Rename: `CommentBanUserDialog` -> `CommentBanCommentDialog`
