+++ src/modules/core/components/Comment/CommentActionsPopover.tsx
@@ -10,7 +10,7 @@ import Button from '~core/Button';
 import Icon from '~core/Icon';
 import { useDialog } from '~core/Dialog';
 import CommentDeleteDialog from '~core/CommentDeleteDialog';
-import CommentBanUserDialog from '~core/CommentBanUserDialog';
+import CommentBanCommentDialog from '~core/CommentBanCommentDialog';
 import { Props as CommentProps } from './Comment';
 
 import styles from './CommentActionsPopover.css';
@@ -46,7 +46,7 @@ const CommentActionsPopover = ({
   fullComment,
 }: Props) => {
   const openDeleteCommentDialog = useDialog(CommentDeleteDialog);
-  const openBanUserDialog = useDialog(CommentBanUserDialog);
+  const openBanUserDialog = useDialog(CommentBanCommentDialog);
 
   const handleDeleteComment = useCallback(
     (undelete = false) =>
