+++ src/modules/core/components/CommentBanCommentDialog/CommentBanCommentDialog.css
