--- src/modules/core/components/CommentBanUserDialog/index.ts
@@ -1,1 +1,0 @@
-export { default } from './CommentBanUserDialog';
