+++ src/modules/dashboard/components/ColonyMembers/ColonyMembers.css
@@ -16,7 +16,7 @@
 
 .rightAside {
   display: flex;
-  padding: 200px 0 20px 50px;
+  padding: 200px 0 20px 100px;
   width: 26.5%;
 }
